refactor: use swr for data fetching in response
